Update todo.js, Create todos.js
todos데이터는 todos.js로 옮겼다.
Map객체를 순회할 때 Array.from(Map.keys())를 사용했다.
Map객체를 만들고 순회할 때 reduce를 사용했다.
결과 메시지를 출력할 때 인라인으로 출력하기 위해 reduce를 사용했다.
