+++ todos.js
@@ -1,0 +1,61 @@
+const todos = [
+    {
+        name: '지구정복하기',
+        category: 'mad science',
+        status: "don't do",
+        id: 1,
+        deadline: "2025-01-01"
+    },
+    {
+        name: '구글 입사',
+        category: 'mission impossible',
+        status: 'doing',
+        id: 2,
+        deadline: "2080-12-31"
+    },
+    {
+        name: '크롱과 뽀로로보기',
+        category: 'chakanil',
+        status: 'todo',
+        id: 3,
+        deadline: '2019-04-17'
+    },
+    {
+        name: '자바스크립트 공부하기',
+        category: 'study',
+        status: 'done',
+        id: 4,
+        deadline: '2019-04-01'
+    },
+    {
+        name: '5만원 다쓰기',
+        category: "현질",
+        status: 'doing',
+        id: 5,
+        deadline: "2019-04-20"
+    },
+    {
+        name: '피자먹기',
+        category: "become a pig",
+        status: 'todo',
+        id: 6,
+        deadline: "2019-06-01"
+    },
+    {
+        name: '창업하기',
+        category: '엄마한테 혼나기',
+        status: 'todo',
+        id: 7,
+        deadline: '2021-01-01'
+
+    },
+    {
+        name: 'kaku와 담배피기',
+        category: '흡연충',
+        status: 'smoking',
+        id: 8,
+        deadline: '2019-04-15'
+    }
+]
+
+module.exports = todos;
